Fuse fmpz_mpoly/ctx_*
