+++ src/fmpz_mpoly/ctx.c
@@ -11,6 +11,17 @@
 
 #include "fmpz_mpoly.h"
 
+void fmpz_mpoly_ctx_clear(fmpz_mpoly_ctx_t ctx)
+{
+    mpoly_ctx_clear(ctx->minfo);
+}
+
+void fmpz_mpoly_ctx_init(fmpz_mpoly_ctx_t ctx,
+                                             slong nvars, const ordering_t ord)
+{
+    mpoly_ctx_init(ctx->minfo, nvars, ord);
+}
+
 void fmpz_mpoly_ctx_init_rand(fmpz_mpoly_ctx_t ctx, flint_rand_t state, slong max_nvars)
 {
     mpoly_ctx_init_rand(ctx->minfo, state, max_nvars);
