--- src/fmpz_mpoly/ctx_init.c
@@ -1,18 +1,0 @@
-/*
-    Copyright (C) 2017 Daniel Schultz
-
-    This file is part of FLINT.
-
-    FLINT is free software: you can redistribute it and/or modify it under
-    the terms of the GNU Lesser General Public License (LGPL) as published
-    by the Free Software Foundation; either version 2.1 of the License, or
-    (at your option) any later version.  See <https://www.gnu.org/licenses/>.
-*/
-
-#include "fmpz_mpoly.h"
-
-void fmpz_mpoly_ctx_init(fmpz_mpoly_ctx_t ctx,
-                                             slong nvars, const ordering_t ord)
-{
-    mpoly_ctx_init(ctx->minfo, nvars, ord);
-}
